Server Folder to keep Organize
I added some Server files. I worked on ServerGame to keep the game and server files seperate. There are a few methods i need adding. The serverGame will initlized the board so the server will need to use the serverGame methods to changed the board. This is a to prohibit direct contact between server and the board as another wall to prevent cheating. I'm currently working on the test cases. I worked out the case to see if the correct color is the expected color from the expected person. I'm using the moves file to check to see if the valid moves(have no idea whats happening in that file tbh) and i'm working on the method that will change the board locations of the pieces.