Finish creating the board in the server game when game starts
When the game starts, the board is inialized. I made a method to output each square in text to make sure the board is being created correctly. I'm leaving it in at the moment to test future tests for moving the pieces around and making sure they pass the checks correctly.